# MAINT, TST, ENH: Simplify flux diagram generation algo, update directional diagram algo to leverage `nx.all_simple_edge_paths()`