* Simplify `_append_reverse_edges()`, `_get_directional_edges()` and
`_flux_edge_conditions()`

* Simplify `_find_unique_edges()`. Performance comparisons
using %timeit showed slight improvement with new algo

* Simplify `_find_unique_uncommon_edges()` and change call
signature such that it takes a graph object and the list
of forward and reverse cycle edges

* Multiple simplifications for `generate_flux_diagrams()`:
  - Generate fresh diagrams instead of making copies
    of a base diagram. This was demonstrated to be
    faster for directional diagram generation.
  - Remove unnecessary `add_nodes_from()` call
  - Simplify node counting
  - Correct call signature for `_find_unique_uncommon_edges()`

* Rename `Test_Flux_Calcs` to `Test_Flux_Diagrams`

* Add 8-state model test case to `test_flux_diagram_counts()`

* Rename `_get_directional_path_edges()` to
`_get_flux_path_edges()` since it is now used
only for flux diagram generation

* Add functions `_collect_sources()` and
`_get_directional_path_edges()` which now
handle the spanning tree pathway finding
for `generate_directional_diagrams()`

* Change `generate_directional_diagrams()`
to use actual diagram objects for generating
the directional diagrams, as is required for
leveraging the NetworkX built-in path finding
function `all_simple_edge_paths()`