@nawtrey 2022-03-11
The code coverage drop is as a result of there being less lines of code overall, which is a net improvement. So that's not a concern.